+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add "Add argument to predicate/non-terminal" refactoring support
 * Add "Reorder predicate/non-terminal arguments" refactoring support
 * Add "Remove argument from predicate/non-terminal" refactoring support
+* Add "Extract protocol" refactoring support
 * Add "Extract to new Logtalk entity" refactoring support
 * Add "Extract to new Logtalk file" refactoring support
 * Add "Replace with include/1 directive" refactoring support